+++ ksql-engine/src/main/java/io/confluent/ksql/ddl/commands/DropTypeFactory.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2019 Confluent Inc.
+ *
+ * Licensed under the Confluent Community License (the "License"); you may not use
+ * this file except in compliance with the License.  You may obtain a copy of the
+ * License at
+ *
+ * http://www.confluent.io/confluent-community-license
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package io.confluent.ksql.ddl.commands;
+
+import io.confluent.ksql.execution.ddl.commands.DropTypeCommand;
+import io.confluent.ksql.parser.DropType;
+
+public class DropTypeFactory {
+  DropTypeFactory() {
+  }
+
+  public DropTypeCommand create(final DropType statement) {
+    return new DropTypeCommand(statement.getTypeName());
+  }
+}
